rv32i-cc: Add clang wrapper script for use with official LLVM prebuilt toolchains
This is for use with builds available from https://releases.llvm.org/download.html
Tested with 15.0.0 and zstd 1.5.6 on macOS Ventura 13.6.9

Note that the toolchain does not come with a libc, so successful linking would require additional
nostdlib-esque flags.  I didn't bother to develop a Logisim converter because LLVM does not support
objcopy's binary format that I have used for this in the past.