cross-cc.sh: Drop C compiler support
C has never been a focus of the microarchitecture course, and this toolchain was of limited utility
because it did not include newlib.  If anyone wants to use something besides assembly for their
project, I can work with them directly to set up Rust or Clang, or build a GCC toolchain.